requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.11.7
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
bleach==5.0.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cryptography==37.0.4
cssselect==1.1.0
decorator==5.1.1
dicttoxml==1.7.4
dill==0.3.5.1
Django==2.2
django-dotenv==1.4.2
docutils==0.19
firebase-admin==5.2.0
frozenlist==1.3.0
future==0.18.2
geocoder==1.38.1
geoip2==4.6.0
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
grpcio==1.47.0
grpcio-status==1.47.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
install==1.3.5
ipinfo==4.2.0
ipinfo-django==1.2.0
isort==5.10.1
JSON-minify==0.3.0
keyring==23.6.0
lazy-object-proxy==1.7.1
lxml==4.9.1
maxminddb==2.2.0
mccabe==0.7.0
MouseInfo==0.1.3
msgpack==1.0.4
multidict==6.0.2
outcome==1.2.0
packaging==21.3
pip-review==1.1.1
pkginfo==1.8.3
platformdirs==2.5.2
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.12.0
pylint==2.14.4
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pyquery==1.4.3
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pytweening==1.0.4
pytz==2022.1
pywin32-ctypes==0.2.0
ratelim==0.1.6
readme-renderer==35.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.4
rsa==4.8
screeninfo==0.8
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
twine==4.0.1
typed-ast==1.5.4
typing==3.7.4.3
ua-parser==0.15.0
uritemplate==4.1.1
urllib3==1.26.10
user-agents==2.2.0
webencodings==0.5.1
wrapt==1.14.1
wsproto==1.1.0
yarl==1.7.2
zipp==3.8.0
